Update 24-game.md
I put in an alternate solution that is similar to the one already there, but uses nested parentheses. It finds quite a few more answers, and I include them below. I could maybe use them to update the tests, but it felt a little presumptuous to just do that; I'm new here!